NgxBootstrapConfirm

+++ my-recipes/src/app/recipes/recipe-detail/recipe-detail.component.ts
@@ -2,36 +2,8 @@ import { Component, OnInit, Input } from '@angular/core';
 import { RecipeService } from 'src/app/shared/recipe.service';
 import { ShoppingListService } from 'src/app/shared/shopping-list.service';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { NgbActiveModal, NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { NgxBootstrapConfirmService } from 'ngx-bootstrap-confirm';
 
-/*
-@Component({
-  selector: 'ngbd-modal-confirm',
-  template: `
-  <div class="modal-header">
-    <h4 class="modal-title" id="modal-title">Profile deletion</h4>
-    <button type="button" class="btn-close" aria-describedby="modal-title" (click)="modal.dismiss('Cross click')">Close</button>
-  </div>
-  <div class="modal-body">
-    <p><strong>Are you sure you want to delete the recipe?</strong></p>
-    <p>All information associated to this recipe will be permanently deleted.
-    <span class="text-danger">This operation can not be undone.</span>
-    </p>
-  </div>
-  <div class="modal-footer">
-    <button type="button" class="btn btn-outline-secondary" (click)="modal.dismiss('cancel click')">Cancel</button>
-    <button type="button" class="btn btn-danger" (click)="modal.close('Ok click')">Ok</button>
-  </div>
-  `
-})
-
-
-export class NgbdModalConfirm {
-  constructor(public modal: NgbActiveModal) {}
-}
-
-*/
 
 @Component({
   selector: 'app-recipe-detail',
@@ -46,7 +18,7 @@ export class RecipeDetailComponent implements OnInit {
     private shoppingList: ShoppingListService,
     private activatedRoute: ActivatedRoute,
     private router: Router,
-    private _modalService: NgbModal
+    private ngxBootstrapConfirmService: NgxBootstrapConfirmService,
   ) { }
 
  
@@ -73,12 +45,32 @@ export class RecipeDetailComponent implements OnInit {
   }
 
   onDeleteRecipe() {
-   /* if(confirm("Are you sure to delete the recipe :"+ this.recipeService.selectedRecipe.name)) {
-      this.recipeService.deleteRecipe(this.recipeService.selectedRecipe.id);
-    }*/
+    /* 
+      if(confirm("Are you sure to delete the recipe :"+ this.recipeService.selectedRecipe.name)) {
+        this.recipeService.deleteRecipe(this.recipeService.selectedRecipe.id);
+      }
+    */
     
-    //this._modalService.open(NgbdModalConfirm)
 
+    console.log("cancellazione effettuata");
+    this.recipeService.deleteRecipe(this.recipeService.selectedRecipe.id);
+
+  }
+
+  action() {
+    let options ={
+      title: 'Sure you want to delete this comment?',
+      confirmLabel: 'Okay',
+      declineLabel: 'Cancel'
+    }
+    this.ngxBootstrapConfirmService.confirm(options).then((res: boolean) => {
+      if (res) {
+        console.log('Okay');
+        this.onDeleteRecipe();
+      } else {
+        console.log('Cancel');
+      }
+    });
   }
 
 }
